change migtrations to mysql

+++ database/migrations/2024_03_05_144449_create_job_vacancies_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('description')->nullable();
             $table->string('additional_information')->nullable();
             $table->datetime('deadline');
-            $table->integer('company_id')->unsigned()->index();
+            $table->bigInteger('company_id')->unsigned()->index();
             $table->foreign('company_id')->references('id')->on('company')->onDelete('cascade');
             $table->timestamps();
         });
